Feat: Add token authentication to protect endpoinst

+++ app/api/posts.py
@@ -1,16 +1,19 @@
 from app.api import bp
 from app.models import Post, User
-from flask import jsonify, request, url_for
+from flask import jsonify, request, url_for, abort
 from app import db
 from app.api.errors import bad_request
+from app.api.auth import token_auth
 
 
 @bp.route('/posts/<int:id>', methods=['GET'])
+@token_auth.login_required
 def get_post(id):
     return jsonify(Post.query.get_or_404(id).to_dict())
 
 
 @bp.route('/posts', methods=['GET'])
+@token_auth.login_required
 def get_posts():
     page = request.args.get('page', 1, type=int)
     per_page = min(request.args.get('per_page', 10, type=int), 100)
@@ -19,6 +22,7 @@ def get_posts():
 
 
 @bp.route('/posts/user/<int:id>', methods=['POST'])
+@token_auth.login_required
 def create_post(id):
     data = request.get_json() or {}
     user = User.query.get_or_404(id)
@@ -34,8 +38,11 @@ def create_post(id):
 
 
 @bp.route('/posts/<int:id>', methods=['PUT'])
+@token_auth.login_required
 def update_post(id):
     post = Post.query.get_or_404(id)
+    if token_auth.current_user().id != post.author.id:
+        abort(403)
     data = request.get_json() or {}
     post.from_dict(data)
     db.session.commit()
@@ -43,8 +50,11 @@ def update_post(id):
 
 
 @bp.route('/posts/<int:id>', methods=['DELETE'])
+@token_auth.login_required
 def delete_post(id):
     post = Post.query.get_or_404(id)
+    if token_auth.current_user().id != post.author.id:
+        abort(403)
     db.session.delete(post)
     db.session.commit()
     page = request.args.get('page', 1, type=int)
